Moving to standalone cli folder
