+++ armory/__main__.py
@@ -23,7 +23,7 @@ from armory.configuration import load_global_config, save_config
 from armory.eval import Evaluator
 import armory.logs
 from armory.logs import log
-from armory.utils.cli import log_current_branch
+from armory.cli.tools import log_current_branch
 from armory.utils.configuration import load_config, load_config_stdin
 from armory.utils.version import to_docker_tag
 import docker
