+++ armory/cli/tools.py
